refactor(porwr): group pages by the surface they belong to
The panel and the management dashboard are two different products
sharing one bundle, and `src/pages/` gave no sign of which page served
which. Every page now sits in exactly one surface directory:

- `pages/sidebar/` — the panel
- `pages/dashboard/` — the full-tab management surfaces
- `pages/extension/` — the entry point, vault login, and 404

The third directory exists because the login, entry, and not-found
surfaces belong to neither of the other two, and leaving them loose at
the top level would have kept the top level meaning "unsorted".

Relative `../` imports in the moved pages resolved to `src/` before the
move and would have resolved to `src/pages/` after it. They are now
aliases, so a page surviving the next move does not depend on how deep
it sits. Where two spellings existed, this takes the one already
dominant in the tree: `components/` and `src/stores/`.

The surface-boundary guard identifies layouts and pages by the name the
SFC compiler derives from the filename. Nothing is renamed here, so the
names it asserts are unchanged and the guard keeps its meaning.

Refs #170
