+++ src/pages/dashboard/AddNewKeyPage.vue
