+++ src/pages/dashboard/ConnectedSitesPage.vue
