+++ src/pages/dashboard/ExtensionSettings.vue
@@ -1,7 +1,7 @@
 <script lang="ts" setup>
 import { ref, onMounted } from 'vue';
-import ThemeSwitch from '../components/ThemeSwitch/Index.vue';
-import useSettingsStore from '../stores/settings-store';
+import ThemeSwitch from 'components/ThemeSwitch/Index.vue';
+import useSettingsStore from 'src/stores/settings-store';
 import { useI18n } from 'vue-i18n';
 import { useQuasar } from 'quasar';
 import { useVaultManagement } from 'src/composables/useVaultManagement';
