+++ src/pages/dashboard/KeyManagementPage.vue
@@ -1,7 +1,7 @@
 <script lang="ts" setup>
 import { computed, onMounted } from 'vue';
 import { useI18n } from 'vue-i18n';
-import useAccountStore from '../stores/account-store';
+import useAccountStore from 'src/stores/account-store';
 import KeyManagementTable from 'components/key-management/KeyManagementTable.vue';
 import SecurityWarning from 'components/SecurityWarning.vue';
