+++ src/pages/dashboard/WalletConnectionsPage.vue
