+++ src/pages/extension/VaultLogin.vue
